+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Road map
 
-- [x] Swap two selections of text.
-- [ ] Hide Command in menu when more or less than two selections made.
+- [ ] ?...
 
 Features that have a checkmark are complete and available for
 download in the
@@ -12,7 +11,13 @@ download in the
 These are the changes to each version that has been released
 on the official Visual Studio extension gallery.
 
+## 1.1
+
+- [X] Hide Command in menu when more or less than two selections made.
+- [X] Added Visual adornment to signal when only 2 spans are selected.
+- [X] Add Options Page to edit Color and Type of text adornment.
+
 ## 1.0
 
 - [x] Initial release
-- [x] Swap two selections of text.
+- [x] Swap two selections of text.